add in login - no added integration
the work requried tro fullyy integrate the users into the app is large.

adding a new user and refactoring everything to incorporate it is substantial and not worth it.

the concept just shows that auth is possible, displaying user data (like the username at the bottom) and then continuing as normal.

adding in user/admin visibility persmission si prrof of abilabty and so on.

this is probably it.